Move to top level type
